requirements.txt
Flask == 1.0.2
SQLAlchemy == 1.3.3
oslo.utils == 3.40.3
oslo.config == 6.8.1
oslo.db == 4.45.0
PyMysql == 0.9.3
validator.py == 1.3.0